requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
fonttools==4.51.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.4
numpy==1.26.4
openai==1.32.0
packaging==23.1
pillow==10.3.0
psycopg2-binary==2.9.7
pydantic==2.7.3
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3
requests==2.31.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.30.1
whitenoise==6.5.0
